+++ src/main/java/woowacourse/auth/config/AuthenticationPrincipalConfig.java
@@ -1,11 +1,9 @@
 package woowacourse.auth.config;
 
 import java.util.List;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
 import woowacourse.auth.application.AuthService;
 import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
