+++ src/main/java/woowacourse/auth/support/AuthorizationExtractor.java
@@ -1,7 +1,6 @@
 package woowacourse.auth.support;
 
 import java.util.Enumeration;
-
 import javax.servlet.http.HttpServletRequest;
 
 public class AuthorizationExtractor {
