+++ src/main/java/woowacourse/auth/ui/AuthControllerAdvice.java
@@ -1,12 +1,10 @@
 package woowacourse.auth.ui;
 
 import java.util.NoSuchElementException;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
 import woowacourse.auth.dto.ExceptionResponse;
 import woowacourse.auth.exception.InvalidTokenException;
 import woowacourse.auth.exception.LoginFailedException;
