+++ src/main/java/woowacourse/shoppingcart/domain/customer/Password.java
@@ -2,7 +2,6 @@ package woowacourse.shoppingcart.domain.customer;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
-
 import woowacourse.shoppingcart.exception.InvalidPasswordFormatException;
 
 public class Password {
