+++ src/main/java/woowacourse/shoppingcart/infrastructure/jdbc/dao/ProductDao.java
@@ -1,24 +1,15 @@
 package woowacourse.shoppingcart.infrastructure.jdbc.dao;
 
-import java.sql.PreparedStatement;
 import java.util.List;
-import java.util.Objects;
-
 import java.util.Optional;
 import javax.sql.DataSource;
 import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
-
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.domain.customer.Customer;
-import woowacourse.shoppingcart.exception.InvalidProductException;
 
 @Repository
 public class ProductDao {
