+++ src/test/java/woowacourse/shoppingcart/domain/customer/NicknameTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
 import woowacourse.shoppingcart.exception.InvalidNicknameFormatException;
 
 class NicknameTest {
