+++ src/test/java/woowacourse/shoppingcart/domain/customer/PasswordTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-
 import woowacourse.shoppingcart.exception.InvalidPasswordFormatException;
 
 class PasswordTest {
