+++ src/test/java/woowacourse/shoppingcart/infrastructure/jdbc/dao/ProductDaoTest.java
@@ -3,7 +3,6 @@ package woowacourse.shoppingcart.infrastructure.jdbc.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-
 import java.util.Optional;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.DisplayName;
@@ -11,10 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
-
 import woowacourse.shoppingcart.domain.Product;
 
 @JdbcTest
